# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: Proman Backend API
  title: Proman Backend
paths:
...
  /api/user/count:
    get:
      consumes:
      - application/json
      operationId: user-count
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get user count
      tags:
      - User
  /api/users:
    get:
      consumes:
      - application/json
      operationId: user-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get list users
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"